rename:api 폴더 이름 변경
