+++ src/apis/studyService.js
@@ -2,18 +2,21 @@
  * 📌 파일 작성 규칙
  * - 담당하신 API 파트에 본인 이름을 작성하고 로직을 완성해주세요.
  * - client(axios 인스턴스)를 사용하여 요청을 보냅니다.
- * - 에러메시지는 constants에 상수로 사용합니다. 
+ * - 에러메시지는 constants에 상수로 사용합니다.
  * - 기능에 따라 추가해야할 에러메시지 상수도 constants에서 작성해주세요.
- * - 각 함수는 비동기(async/await)로 작성하며, 데이터 반환 형식은 response.data를 기본으로 합니다.
+ * - 각 함수는 비동기(async/await)로 작성하며, 데이터 반환은 (response.data) client.js에서 다루고 있어요.
  */
 
-import { STUDY_ERROR_MESSAGES } from '../constants/error';
 import client from './client';
-
 
 // 1. 스터디 목록 조회 (GET /)
 // 담당: 000
-export const getStudies = async ({ page = 1, pageSize = 10, orderBy = 'recent', keyword }) => {
+export const getStudies = async ({
+  page = 1,
+  pageSize = 10,
+  orderBy = 'recent',
+  keyword,
+}) => {
   try {
     const params = {
       page,
@@ -89,7 +92,13 @@ export const getStudyEmojis = async (studyId) => {
 
 // 6. 스터디 생성 (POST /)
 // 담당: 강에스더
-export const createStudy = async ({ nickname, title, introduction, background, password }) => {
+export const createStudy = async ({
+  nickname,
+  title,
+  introduction,
+  background,
+  password,
+}) => {
   try {
     const response = await client.post('/studies', {
       nickname,
@@ -153,7 +162,9 @@ export const recordFocusTime = async (studyId, focusData) => {
 // 담당: 안예진
 export const verifyStudyPassword = async (studyId, password) => {
   try {
-    const response = await client.post(`/studies/${studyId}/password/verify`, { password });
+    const response = await client.post(`/studies/${studyId}/password/verify`, {
+      password,
+    });
     const data = response.data;
     console.log(data);
     return data;
@@ -179,12 +190,4 @@ export const patchStudy = async (studyId, updateData) => {
 
 // 12. 스터디 삭제 (DELETE /:studyId)
 // 담당: 안예진
-export const deleteStudy = async (studyId) => {
-  try {
-    const response = await client.delete(`/studies/${studyId}`);
-    return response.status;
-  } catch (error) {
-    console.log(`[deleteStudy] - 에러발생: ${error.message}`);
-    throw error;
-  }
-};
+export const deleteStudy = async (studyId, password) => {};
