+++ src/components/HabitsTable/HabitsTable.module.css
@@ -25,7 +25,7 @@
   position: sticky;
   top: 0;
   background-color: white;
-  z-index: 10;
+  z-index: 1;
 }
 
 .habitNameCell {
