+++ src/hooks/todayFocus/useFocusTimer.hook.js
@@ -1,5 +1,5 @@
 import { useTimerAction } from './useTimerAction.hook';
-import { settlePointsApi } from '../../apis/focusTimer.api.js';
+import { settlePointsApi } from '../../apis/focusTimerService.js';
 import { INITIAL_SECONDS } from '../../constants/time.js';
 import styles from '../../components/focusTimer.module.css';
 // import { showToast } from '../../utils/toast.util.js';
